compatible to ffmpeg tag n3.x
